Made the Ticker and Exchange input Dynamic by Using API GetStockTickerExchangesWithoutID, Modified the StockTickerExchange Model, Modified the Primary Key of Candel Model,

+++ Migrations/20241205045000_InitialCreate.cs
@@ -15,7 +15,7 @@ namespace StockLogger.Migrations
                 name: "Candel",
                 columns: table => new
                 {
-                    Id = table.Column<int>(type: "int", nullable: false)
+                    Id = table.Column<long>(type: "bigint", nullable: false)
                         .Annotation("SqlServer:Identity", "1, 1"),
                     StartPrice = table.Column<decimal>(type: "decimal(18,2)", nullable: false),
                     HighestPrice = table.Column<decimal>(type: "decimal(18,2)", nullable: false),
@@ -38,7 +38,7 @@ namespace StockLogger.Migrations
                 name: "StockTickerExchanges",
                 columns: table => new
                 {
-                    Id = table.Column<int>(type: "int", nullable: false)
+                    Id = table.Column<long>(type: "bigint", nullable: false)
                         .Annotation("SqlServer:Identity", "1, 1"),
                     Ticker = table.Column<string>(type: "nvarchar(max)", nullable: false),
                     Exchange = table.Column<string>(type: "nvarchar(max)", nullable: false)
