+++ Models/Candel/Candel.cs
@@ -5,7 +5,7 @@ namespace StockLogger.Models.Candel
     public class Candel
     {
         [Key]
-        public int Id { get; set; }  // Primary key
+        public long Id { get; set; }  // Primary key
 
         // Core properties
         public decimal StartPrice { get; set; }
